pyproject.toml
[project]
name = "mslu-schedule-parser"
version = "0.1.0"
description = ""
authors = [
    {name = "stanislav",email = "s.markovtsev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "ipython (>=9.6.0,<10.0.0)",
    "google-api-python-client (>=2.185.0,<3.0.0)",
    "google-auth (>=2.41.1,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
